컴파일 경고 제거

+++ src/ThruthGameEngine/Scene.cpp
@@ -65,7 +65,7 @@ void Truth::Scene::AddEntity(std::shared_ptr<Entity> _p)
 /// <param name="_p"></param>
 void Truth::Scene::CreateEntity(std::shared_ptr<Entity> _p)
 {
-	_p->m_index = m_entities.size();
+	_p->m_index = static_cast<int32>(m_entities.size());
 	m_awakedEntity.push(_p);
 	_p->m_isAdded = false;
 }
